+++ python/sopActor/Commands/SopCmd.py
@@ -679,8 +679,23 @@ class SopCmd(object):
             if finish:
                 cmd.fail('text="Some lamps failed to turn off"')
 
+    def setRefractionBalance(self, cmd, subSystem, doBypass, sopState,
+                             plateType, surveyMode):
+        """Sets the refraction balance."""
+
+        cmdStr = 'setRefractionBalance plateType="{0}" surveyMode="{1}"'.format(
+            plateType, surveyMode)
+        cmdVar = sopState.actor.cmdr.call(actor='guider', forUserCmd=cmd, cmdStr=cmdStr)
+        if cmdVar.didFail:
+            cmd.fail('text="Failed to set guider refraction '
+                     'balance for bypass {0} {1}'.format(subSystem, doBypass))
+            return False
+
+        return True
+
     def bypass(self, cmd):
         """Ignore errors in a subsystem, or force a system to be in a given state."""
+
         subSystems = cmd.cmd.keywords['subSystem'].values
         doBypass = False if 'clear' in cmd.cmd.keywords else True
 
@@ -688,10 +703,12 @@ class SopCmd(object):
         bypass = myGlobals.bypass
 
         for subSystem in subSystems:
+
             if bypass.set(subSystem, doBypass) is None:
-                cmd.fail(
-                    'text="{} is not a recognised and bypassable subSystem"'.format(subSystem))
+                cmd.fail('text="{} is not a recognised and '
+                         'bypassable subSystem"'.format(subSystem))
                 return
+
             if bypass.is_cart_bypass(subSystem):
                 self.updateCartridge(
                     sopState.cartridge,
@@ -699,16 +716,24 @@ class SopCmd(object):
                     sopState.surveyModeName,
                     status=False,
                     bypassed=True)
-                cmdStr = 'setRefractionBalance plateType="{0}" surveyMode="{1}"'.format(
-                    *sopState.surveyText)
-                cmdVar = sopState.actor.cmdr.call(actor='guider', forUserCmd=cmd, cmdStr=cmdStr)
-                if cmdVar.didFail:
-                    cmd.fail(
-                        'text="Failed to set guider refraction balance for bypass {0} {1}'.format(
-                            subSystem, doBypass))
+                if not self.setRefractionBalance(cmd, subSystem, doBypass,
+                                                 sopState, *sopState.surveyText):
                     return
+
             if bypass.is_gang_bypass(subSystem):
                 cmd.warn('text="gang bypassed: {}"'.format(sopState.apogeeGang.getPos()))
+
+            if subSystem == 'noBOSS':
+                if not self.setRefractionBalance(cmd, sopState, subSystem,
+                                                 doBypass, sopState, 'BHM&MWM',
+                                                 'MWM lead'):
+                    return
+
+            if subSystem == 'noAPOGEE':
+                if not self.setRefractionBalance(cmd, sopState, subSystem,
+                                                 doBypass, sopState, 'BHM&MWM',
+                                                 'BHM lead'):
+                    return
 
         self.status(cmd, threads=False)